requirements.txt

streamlit
pandas
googlemaps
pydeck
openpyxl
altair
sqlite-utils
python-dotenv
